arm: -march for Cortex A7, A12 and A15 is armv7ve
For Cortex A7, A12 and A15 setting -march=armv7-a will generate this
type of warning by default:

warning: switch -mcpu=cortex-a15 conflicts with -march=armv7-a switch

This is because these cores have an instruction set which is based on
the armv7-a set, but adds some extensions on top of them. This new set
is called 'armv7ve' and is the default for A7, A12 and A15:

http://gcc.gnu.org/viewcvs?rev=207237&root=gcc&view=rev

This way we make sure we don't force the warning anymore.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>